+++ test/dotnet-watch.Tests/HotReload/RuntimeProcessLauncherTests.cs
@@ -1,9 +1,6 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System.Runtime.CompilerServices;
-using Microsoft.Extensions.Logging;
-
 namespace Microsoft.DotNet.Watch.UnitTests;
 
 public class RuntimeProcessLauncherTests(ITestOutputHelper logger) : DotNetWatchTestBase(logger)
@@ -407,370 +404,5 @@ public class RuntimeProcessLauncherTests(ITestOutputHelper logger) : DotNetWatch
 
         Log("Waiting for verbose rude edit reported ...");
         await applyUpdateVerbose.WaitAsync(w.ShutdownSource.Token);
-    }
-
-    public enum DirectoryKind
-    {
-        Ordinary,
-        Hidden,
-        Bin,
-        Obj,
-    }
-
-    [Theory]
-    [CombinatorialData]
-    public async Task IgnoredChange(bool isExisting, bool isIncluded, DirectoryKind directoryKind)
-    {
-        var testAsset = CopyTestAsset("WatchNoDepsApp", [isExisting, isIncluded, directoryKind]);
-
-        var workingDirectory = testAsset.Path;
-        string dir;
-
-        switch (directoryKind)
-        {
-            case DirectoryKind.Bin:
-                dir = Path.Combine(workingDirectory, "bin", "Debug", ToolsetInfo.CurrentTargetFramework);
-                break;
-
-            case DirectoryKind.Obj:
-                dir = Path.Combine(workingDirectory, "obj", "Debug", ToolsetInfo.CurrentTargetFramework);
-                break;
-
-            case DirectoryKind.Hidden:
-                dir = Path.Combine(workingDirectory, ".dir");
-                break;
-
-            default:
-                dir = workingDirectory;
-                break;
-        }
-
-        var extension = isIncluded ? ".cs" : ".txt";
-
-        Directory.CreateDirectory(dir);
-
-        var path = Path.Combine(dir, "File" + extension);
-
-        if (isExisting)
-        {
-            File.WriteAllText(path, "class C { int F() => 1; }");
-
-            if (isIncluded && directoryKind is DirectoryKind.Bin or DirectoryKind.Obj or DirectoryKind.Hidden)
-            {
-                var project = Path.Combine(workingDirectory, "WatchNoDepsApp.csproj");
-                File.WriteAllText(project, File.ReadAllText(project).Replace(
-                    "<!-- add item -->",
-                    $"""
-                    <Compile Include="{path}"/>
-                    """));
-            }
-        }
-
-        await using var w = CreateInProcWatcher(testAsset, ["--no-exit"], workingDirectory);
-
-        var waitingForChanges = w.Reporter.RegisterSemaphore(MessageDescriptor.WaitingForChanges);
-        var changeHandled = w.Reporter.RegisterSemaphore(MessageDescriptor.ManagedCodeChangesApplied);
-        var ignoringChangeInHiddenDirectory = w.Reporter.RegisterSemaphore(MessageDescriptor.IgnoringChangeInHiddenDirectory);
-        var ignoringChangeInExcludedFile = w.Reporter.RegisterSemaphore(MessageDescriptor.IgnoringChangeInExcludedFile);
-        var fileAdditionTriggeredReEvaluation = w.Reporter.RegisterSemaphore(MessageDescriptor.FileAdditionTriggeredReEvaluation);
-        var reEvaluationCompleted = w.Reporter.RegisterSemaphore(MessageDescriptor.ReEvaluationCompleted);
-        var noHotReloadChangesToApply = w.Reporter.RegisterSemaphore(MessageDescriptor.NoManagedCodeChangesToApply);
-
-        w.Start();
-
-        Log("Waiting for changes...");
-        await waitingForChanges.WaitAsync(w.ShutdownSource.Token);
-        
-        UpdateSourceFile(path, "class C { int F() => 2; }");
-
-        switch ((isExisting, isIncluded, directoryKind))
-        {
-            case (isExisting: true, isIncluded: true, directoryKind: _):
-                Log("Waiting for changed handled ...");
-                await changeHandled.WaitAsync(w.ShutdownSource.Token);
-                break;
-
-            case (isExisting: true, isIncluded: false, directoryKind: DirectoryKind.Ordinary):
-                Log("Waiting for no hot reload changes to apply ...");
-                await noHotReloadChangesToApply.WaitAsync(w.ShutdownSource.Token);
-                break;
-
-            case (isExisting: false, isIncluded: _, directoryKind: DirectoryKind.Ordinary):
-                Log("Waiting for file addition re-evalutation ...");
-                await fileAdditionTriggeredReEvaluation.WaitAsync(w.ShutdownSource.Token);
-                Log("Waiting for re-evalutation to complete ...");
-                await reEvaluationCompleted.WaitAsync(w.ShutdownSource.Token);
-                break;
-
-            case (isExisting: _, isIncluded: _, directoryKind: DirectoryKind.Hidden):
-                Log("Waiting for ignored change in hidden dir ...");
-                await ignoringChangeInHiddenDirectory.WaitAsync(w.ShutdownSource.Token);
-                break;
-
-            case (isExisting: _, isIncluded: _, directoryKind: DirectoryKind.Bin or DirectoryKind.Obj):
-                Log("Waiting for ignored change in output dir ...");
-                await ignoringChangeInExcludedFile.WaitAsync(w.ShutdownSource.Token);
-                break;
-
-            default:
-                throw new InvalidOperationException();
-        }
-    }
-
-    [Fact]
-    public async Task CtrlR_RestartsBuild()
-    {
-        var testAsset = TestAssets.CopyTestAsset("WatchHotReloadApp")
-            .WithSource();
-
-        await using var w = CreateInProcWatcher(testAsset, []);
-
-        var buildCounter = 0;
-
-        w.Reporter.RegisterAction(MessageDescriptor.Building, () =>
-        {
-            if (Interlocked.Increment(ref buildCounter) == 1)
-            {
-                w.Console.PressKey(new ConsoleKeyInfo('R', ConsoleKey.R, shift: false, alt: false, control: true));
-            }
-        });
-
-        var restarting = w.Reporter.RegisterSemaphore(MessageDescriptor.Restarting);
-
-        // Iteration #1 build should be canceled, iteration #2 should build and launch the app.
-        var hasExpectedOutput = w.CreateCompletionSource();
-        w.Reporter.OnProcessOutput += line =>
-        {
-            Assert.DoesNotContain("DOTNET_WATCH_ITERATION = 1", line.Content);
-
-            if (line.Content.Contains("DOTNET_WATCH_ITERATION = 2"))
-            {
-                hasExpectedOutput.TrySetResult();
-            }
-        };
-
-        w.Start();
-
-        // 🔄 Restarting
-        await restarting.WaitAsync(w.ShutdownSource.Token);
-
-        // DOTNET_WATCH_ITERATION = 2
-        await hasExpectedOutput.Task;
-
-        Assert.Equal(2, buildCounter);
-    }
-
-    [Fact]
-    public async Task CtrlR_CancelsWaitForFileChange()
-    {
-        var testAsset = TestAssets.CopyTestAsset("WatchHotReloadApp")
-            .WithSource();
-
-        var programFilePath = Path.Combine(testAsset.Path, "Program.cs");
-
-        File.WriteAllText(programFilePath, """
-            System.Console.WriteLine("<Started>");
-            """);
-
-        await using var w = CreateInProcWatcher(testAsset, []);
-
-        w.Reporter.RegisterAction(MessageDescriptor.WaitingForFileChangeBeforeRestarting, () =>
-        {
-            w.Console.PressKey(new ConsoleKeyInfo('R', ConsoleKey.R, shift: false, alt: false, control: true));
-        });
-
-        var buildCounter = 0;
-        w.Reporter.RegisterAction(MessageDescriptor.Building, () => Interlocked.Increment(ref buildCounter));
-
-        var counter = 0;
-        var hasExpectedOutput = w.CreateCompletionSource();
-        w.Reporter.OnProcessOutput += line =>
-        {
-            if (line.Content.Contains("<Started>") && Interlocked.Increment(ref counter) == 2)
-            {
-                hasExpectedOutput.TrySetResult();
-            }
-        };
-
-        w.Start();
-
-        await hasExpectedOutput.Task;
-
-        Assert.Equal(2, buildCounter);
-    }
-
-    [Fact]
-    public async Task ProjectAndSourceFileChange()
-    {
-        var testAsset = CopyTestAsset("WatchHotReloadApp");
-
-        var workingDirectory = testAsset.Path;
-        var projectPath = Path.Combine(testAsset.Path, "WatchHotReloadApp.csproj");
-        var programPath = Path.Combine(testAsset.Path, "Program.cs");
-
-        await using var w = CreateInProcWatcher(testAsset, [], workingDirectory);
-
-        var fileChangesCompleted = w.CreateCompletionSource();
-        w.Watcher.Test_FileChangesCompletedTask = fileChangesCompleted.Task;
-
-        var waitingForChanges = w.Reporter.RegisterSemaphore(MessageDescriptor.WaitingForChanges);
-
-        var changeHandled = w.Reporter.RegisterSemaphore(MessageDescriptor.ManagedCodeChangesApplied);
-
-        var hasUpdatedOutput = w.CreateCompletionSource();
-        w.Reporter.OnProcessOutput += line =>
-        {
-            if (line.Content.Contains("System.Xml.Linq.XDocument"))
-            {
-                hasUpdatedOutput.TrySetResult();
-            }
-        };
-
-        w.Start();
-
-        Log("Waiting for changes...");
-        await waitingForChanges.WaitAsync(w.ShutdownSource.Token);
-
-        // change the project and source files at the same time:
-
-        UpdateSourceFile(programPath, src => src.Replace("""Console.WriteLine(".");""", """Console.WriteLine(typeof(XDocument));"""));
-        UpdateSourceFile(projectPath, src => src.Replace("<!-- items placeholder -->", """<Using Include="System.Xml.Linq"/>"""));
-
-        // done updating files:
-        fileChangesCompleted.TrySetResult();
-
-        Log("Waiting for change handled ...");
-        await changeHandled.WaitAsync(w.ShutdownSource.Token);
-
-        Log("Waiting for output 'System.Xml.Linq.XDocument'...");
-        await hasUpdatedOutput.Task;
-    }
-
-    [Fact]
-    public async Task ProjectAndSourceFileChange_AddProjectReference()
-    {
-        var testAsset = TestAssets.CopyTestAsset("WatchAppWithProjectDeps")
-            .WithSource()
-            .WithProjectChanges(project =>
-            {
-                foreach (var r in project.Root!.Descendants().Where(e => e.Name.LocalName == "ProjectReference").ToArray())
-                {
-                    r.Remove();
-                }
-            });
-
-        var appProjDir = Path.Combine(testAsset.Path, "AppWithDeps");
-        var appProjFile = Path.Combine(appProjDir, "App.WithDeps.csproj");
-        var appFile = Path.Combine(appProjDir, "Program.cs");
-
-        UpdateSourceFile(appFile, code => code.Replace("Lib.Print();", "// Lib.Print();"));
-
-        await using var w = CreateInProcWatcher(testAsset, [], appProjDir);
-
-        var fileChangesCompleted = w.CreateCompletionSource();
-        w.Watcher.Test_FileChangesCompletedTask = fileChangesCompleted.Task;
-
-        var waitingForChanges = w.Reporter.RegisterSemaphore(MessageDescriptor.WaitingForChanges);
-        var projectChangeTriggeredReEvaluation = w.Reporter.RegisterSemaphore(MessageDescriptor.ProjectChangeTriggeredReEvaluation);
-        var projectsRebuilt = w.Reporter.RegisterSemaphore(MessageDescriptor.ProjectsRebuilt);
-        var projectDependenciesDeployed = w.Reporter.RegisterSemaphore(MessageDescriptor.ProjectDependenciesDeployed);
-        var managedCodeChangesApplied = w.Reporter.RegisterSemaphore(MessageDescriptor.ManagedCodeChangesApplied);
-
-        var hasUpdatedOutput = w.CreateCompletionSource();
-        w.Reporter.OnProcessOutput += line =>
-        {
-            if (line.Content.Contains("<Lib>"))
-            {
-                hasUpdatedOutput.TrySetResult();
-            }
-        };
-
-        w.Start();
-
-        Log("Waiting for changes...");
-        await waitingForChanges.WaitAsync(w.ShutdownSource.Token);
-
-        // change the project and source files at the same time:
-
-        UpdateSourceFile(appProjFile, src => src.Replace("""
-            <ItemGroup />
-            """, """
-            <ItemGroup>
-                <ProjectReference Include="..\Dependency\Dependency.csproj" />
-            </ItemGroup>
-            """));
-
-        UpdateSourceFile(appFile, code => code.Replace("// Lib.Print();", "Lib.Print();"));
-
-        // done updating files:
-        fileChangesCompleted.TrySetResult();
-
-        Log("Waiting for output '<Lib>'...");
-        await hasUpdatedOutput.Task;
-
-        AssertEx.ContainsSubstring("Resolving 'Dependency, Version=1.0.0.0'", w.Reporter.ProcessOutput);
-
-        Assert.Equal(1, projectChangeTriggeredReEvaluation.CurrentCount);
-        Assert.Equal(1, projectsRebuilt.CurrentCount);
-        Assert.Equal(1, projectDependenciesDeployed.CurrentCount);
-        Assert.Equal(1, managedCodeChangesApplied.CurrentCount);
-    }
-
-    [Fact]
-    public async Task ProjectAndSourceFileChange_AddPackageReference()
-    {
-        var testAsset = TestAssets.CopyTestAsset("WatchHotReloadApp")
-            .WithSource();
-
-        var projFilePath = Path.Combine(testAsset.Path, "WatchHotReloadApp.csproj");
-        var programFilePath = Path.Combine(testAsset.Path, "Program.cs");
-
-        await using var w = CreateInProcWatcher(testAsset, []);
-
-        var fileChangesCompleted = w.CreateCompletionSource();
-        w.Watcher.Test_FileChangesCompletedTask = fileChangesCompleted.Task;
-
-        var waitingForChanges = w.Reporter.RegisterSemaphore(MessageDescriptor.WaitingForChanges);
-        var projectChangeTriggeredReEvaluation = w.Reporter.RegisterSemaphore(MessageDescriptor.ProjectChangeTriggeredReEvaluation);
-        var projectsRebuilt = w.Reporter.RegisterSemaphore(MessageDescriptor.ProjectsRebuilt);
-        var projectDependenciesDeployed = w.Reporter.RegisterSemaphore(MessageDescriptor.ProjectDependenciesDeployed);
-        var managedCodeChangesApplied = w.Reporter.RegisterSemaphore(MessageDescriptor.ManagedCodeChangesApplied);
-
-        var hasUpdatedOutput = w.CreateCompletionSource();
-        w.Reporter.OnProcessOutput += line =>
-        {
-            if (line.Content.Contains("Newtonsoft.Json.Linq.JToken"))
-            {
-                hasUpdatedOutput.TrySetResult();
-            }
-        };
-
-        w.Start();
-
-        Log("Waiting for changes...");
-        await waitingForChanges.WaitAsync(w.ShutdownSource.Token);
-
-        // change the project and source files at the same time:
-
-        UpdateSourceFile(projFilePath, source => source.Replace("""
-            <!-- items placeholder -->
-            """, """
-            <PackageReference Include="Newtonsoft.Json" Version="13.0.3"/>
-            """));
-
-        UpdateSourceFile(programFilePath, source => source.Replace("Console.WriteLine(\".\");", "Console.WriteLine(typeof(Newtonsoft.Json.Linq.JToken));"));
-
-        // done updating files:
-        fileChangesCompleted.TrySetResult();
-
-        Log("Waiting for output 'Newtonsoft.Json.Linq.JToken'...");
-        await hasUpdatedOutput.Task;
-
-        AssertEx.ContainsSubstring("Resolving 'Newtonsoft.Json, Version=13.0.0.0'", w.Reporter.ProcessOutput);
-
-        Assert.Equal(1, projectChangeTriggeredReEvaluation.CurrentCount);
-        Assert.Equal(0, projectsRebuilt.CurrentCount);
-        Assert.Equal(1, projectDependenciesDeployed.CurrentCount);
-        Assert.Equal(1, managedCodeChangesApplied.CurrentCount);
     }
 }
